pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rlabs-agentguard"
version = "0.1.0"
description = "A quality-assurance engine for LLM-generated code"
readme = "README.MD"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "AgentGuard Team" }]
keywords = ["llm", "code-generation", "quality-assurance", "ai", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Quality Assurance",
    "Typing :: Typed",
]

dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-PyYAML",
    "httpx>=0.27.0",
]
server = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "sse-starlette>=2.0.0",
]
mcp = ["mcp>=1.0.0"]
litellm = ["litellm>=1.50.0"]
google = ["google-genai>=1.0.0"]
platform = ["httpx>=0.27.0"]
all = ["rlabs-agentguard[server,mcp,litellm,google,platform]"]

[project.scripts]
agentguard = "agentguard.cli:main"

# --- Hatch Build ---
[tool.hatch.build.targets.wheel]
packages = ["agentguard"]

[project.urls]
Homepage = "https://github.com/rlabs-cl/agentguard-lib"
Documentation = "https://github.com/rlabs-cl/agentguard-lib#readme"
Repository = "https://github.com/rlabs-cl/agentguard-lib"
Issues = "https://github.com/rlabs-cl/agentguard-lib/issues"
Changelog = "https://github.com/rlabs-cl/agentguard-lib/releases"

# --- Ruff ---
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "TCH"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["agentguard"]

# --- Mypy ---
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["anthropic.*", "openai.*", "litellm.*", "google.*", "mcp.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "agentguard.mcp.server",
    "agentguard.mcp.tools",
    "agentguard.server.routes",
    "agentguard.server.app",
    "agentguard.server.auth",
]
disallow_untyped_decorators = false

# --- Pytest ---
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests that call real LLM APIs",
]

# --- Coverage ---
[tool.coverage.run]
source = ["agentguard"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_empty = true
